variation for float32 instead of double

+++ dtaidistance/lib/DTAIDistanceC/DTAIDistanceC/dd_globals.h
@@ -16,7 +16,7 @@
 
 
 /* The sequence type type can be customized by changing the typedef. */
-typedef double seq_t;
+typedef float seq_t;
 
 /*! The index type
